+++ tests/test_convert.py
@@ -13,7 +13,7 @@ def test_basic(tmp_path, monkeypatch):
     path = (enex_root / "notebook01.enex").absolute()
     monkeypatch.chdir(tmp_path)
     converter = Converter()
-    converter.convert(enex=path, sink=FileSystemSink.legacy_root_from_enex(path))
+    converter.convert(enex=path, sink=FileSystemSink())
 
     generated_files = [p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file()]
     assert len(generated_files) == 1
@@ -46,7 +46,7 @@ def test_frontmatter(tmp_path, monkeypatch):
     path = (enex_root / "notebook01.enex").absolute()
     monkeypatch.chdir(tmp_path)
     converter = Converter(front_matter=True)
-    converter.convert(enex=path, sink=FileSystemSink.legacy_root_from_enex(path))
+    converter.convert(enex=path, sink=FileSystemSink())
 
     generated_files = [p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file()]
     assert len(generated_files) == 1
@@ -78,7 +78,7 @@ def test_nested_lists(tmp_path, monkeypatch):
     path = (enex_root / "notebook02.enex").absolute()
     monkeypatch.chdir(tmp_path)
     converter = Converter()
-    converter.convert(enex=path, sink=FileSystemSink.legacy_root_from_enex(path))
+    converter.convert(enex=path, sink=FileSystemSink())
 
     generated_files = [p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file()]
     assert len(generated_files) == 1
@@ -122,11 +122,11 @@ def test_nested_lists(tmp_path, monkeypatch):
 
 
 @pytest.mark.parametrize("enex", ["notebook03.enex", "notebook03-2.enex"])
-def test_attachment(tmp_path, monkeypatch, enex):
+def test_attachment_default_paths(tmp_path, monkeypatch, enex):
     path = (enex_root / enex).absolute()
     monkeypatch.chdir(tmp_path)
     converter = Converter()
-    converter.convert(enex=path, sink=FileSystemSink.legacy_root_from_enex(path))
+    converter.convert(enex=path, sink=FileSystemSink())
 
     generated_files = [p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file()]
     assert len(generated_files) == 2
@@ -154,6 +154,125 @@ def test_attachment(tmp_path, monkeypatch, enex):
         """
     )
     assert png_path.read_bytes()[:8] == b"\x89PNG\r\n\x1a\n"
+
+
+def test_custom_paths_basic(tmp_path, monkeypatch):
+    path = (enex_root / "notebook01.enex").absolute()
+    monkeypatch.chdir(tmp_path)
+    converter = Converter()
+    converter.convert(
+        enex=path,
+        sink=FileSystemSink(
+            root=tmp_path / "go" / "here",
+            note_path_template="{enex}/{created:%Y}/{created:%Y%m%d}-{title}.md",
+        ),
+    )
+
+    generated_files = [p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file()]
+    assert generated_files == [
+        Path("go/here/notebook01/2023/20230709-The_title.md"),
+    ]
+    md_path = generated_files[0]
+    md_content = md_path.read_text()
+    assert md_content == textwrap.dedent(
+        """\
+        # The title
+
+        ## Note metadata
+
+        - Title: The title
+        - Author: John Doe
+        - Created: 2023-07-09T18:42:04+00:00
+        - Updated: 2023-07-09T18:43:22+00:00
+
+        ## Note Content
+
+        Things to buy:
+
+          * apple
+          * banana
+          * chocolate
+        """
+    )
+
+
+def test_custom_paths_auto_attachments(tmp_path, monkeypatch):
+    path = (enex_root / "notebook03.enex").absolute()
+    monkeypatch.chdir(tmp_path)
+    converter = Converter()
+    converter.convert(
+        enex=path,
+        sink=FileSystemSink(
+            root="dump",
+            note_path_template="{enex}/{created:%Y}/{created:%Y%m%d}-{title}.md",
+        ),
+    )
+
+    generated_files = sorted(p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file())
+    assert generated_files == [
+        Path("dump/notebook03/2023/20230712-Fa_fa_fa.md"),
+        Path("dump/notebook03/2023/20230712-Fa_fa_fa_attachments/rckrll.png"),
+    ]
+    assert generated_files[0].read_text() == textwrap.dedent(
+        """\
+        # Fa fa fa
+
+        ## Note metadata
+
+        - Title: Fa fa fa
+        - Author: John Doe
+        - Created: 2023-07-12T20:16:08+00:00
+        - Updated: 2023-07-12T20:18:38+00:00
+
+        ## Note Content
+
+        lo lo lo
+
+        ![rckrll.png](20230712-Fa_fa_fa_attachments/rckrll.png)
+        la la la
+        """
+    )
+    assert generated_files[1].read_bytes()[:8] == b"\x89PNG\r\n\x1a\n"
+
+
+def test_custom_paths(tmp_path, monkeypatch):
+    path = (enex_root / "notebook03.enex").absolute()
+    monkeypatch.chdir(tmp_path)
+    converter = Converter()
+    converter.convert(
+        enex=path,
+        sink=FileSystemSink(
+            root="dump",
+            note_path_template="{enex}/{created:%Y}/{created:%Y%m%d}-{title}.md",
+            attachments_path_template="{enex}/_resources/{created:%Y}/{created:%Y%m%d}-{title}",
+        ),
+    )
+
+    generated_files = sorted(p.relative_to(tmp_path) for p in tmp_path.glob("**/*") if p.is_file())
+    assert generated_files == [
+        Path("dump/notebook03/2023/20230712-Fa_fa_fa.md"),
+        Path("dump/notebook03/_resources/2023/20230712-Fa_fa_fa/rckrll.png"),
+    ]
+    assert generated_files[0].read_text() == textwrap.dedent(
+        """\
+        # Fa fa fa
+
+        ## Note metadata
+
+        - Title: Fa fa fa
+        - Author: John Doe
+        - Created: 2023-07-12T20:16:08+00:00
+        - Updated: 2023-07-12T20:18:38+00:00
+
+        ## Note Content
+
+        lo lo lo
+
+        ![rckrll.png](../_resources/2023/20230712-Fa_fa_fa/rckrll.png)
+        la la la
+        """
+    )
+    assert generated_files[1].read_bytes()[:8] == b"\x89PNG\r\n\x1a\n"
 
 
 class TestEnexParser: