+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add option to add note metadata in "frontmatter" style instead of Markdown section
 - Started with basic test suite
 - Added GitHub actions for running test suite and linting checks
+- Add "path template" options to customize output tree structure (both for notes and their attachments)
 
 ### Changed